Extract performance calculation evidence boundary
Evidence:
- python -m ruff check focused touched files: passed
- python -m ruff format --check focused touched files: passed
- python -m pytest tests/unit/test_performance_workspace_evidence.py tests/unit/test_refactor_quality_thresholds.py tests/unit/test_performance_workspace_service.py -q: 59 passed
- make check: passed; 1,190 unit/contract tests
- make ci: passed; 209 integration tests, 1,399 coverage tests, 94.24% total coverage, pip-audit no known vulnerabilities after governed PYSEC-2026-161 exception
- Sync-RepoWikis.ps1 -CheckOnly -Repository lotus-gateway: expected drift on Architecture.md and Validation-and-CI.md before merge/publish
